Refactor city generation
