+++ Building.cs
@@ -6,5 +6,11 @@ namespace StrategyGame
     {
         public Nts.Polygon Footprint { get; set; }
         public LandUseType LandUse { get; set; }
+
+        // Fields used by the BuildingRefiner
+        public int Level { get; set; }
+        public int PopulationCapacity { get; set; }
+        public double EconomicOutput { get; set; }
+        public double PollutionOutput { get; set; }
     }
 }
